Split public interface, only exports factory now
