+++ src/CMakeLists.txt
@@ -6,6 +6,7 @@ set(SRC_FILES ${SRC_FILES}
     ${CMAKE_CURRENT_SOURCE_DIR}/AttributeName.hpp
     ${CMAKE_CURRENT_SOURCE_DIR}/AttributeTraits.cpp
     ${CMAKE_CURRENT_SOURCE_DIR}/HdbppTimescaleDb.cpp
+    ${CMAKE_CURRENT_SOURCE_DIR}/HdbppTimescaleDbImpl.cpp
     ${CMAKE_CURRENT_SOURCE_DIR}/LibUtils.cpp
     ${CMAKE_CURRENT_SOURCE_DIR}/DbConnection.cpp
     ${CMAKE_CURRENT_SOURCE_DIR}/QueryBuilder.cpp
